+++ run_test.sh
@@ -1,0 +1,2 @@
+pip install . -U
+python -m pytest -v tests/
